backend/requirements.txt
flask
flask-cors
python-dotenv
google-generativeai
PyMuPDF
supabase